+++ Source/MainSpaConBridgeComponent.cpp
@@ -43,11 +43,12 @@ MainSpaConBridgeComponent::MainSpaConBridgeComponent()
 MainSpaConBridgeComponent::MainSpaConBridgeComponent(std::function<void(DbLookAndFeelBase::LookAndFeelType)> lafUpdateCallback)
     : onUpdateLookAndFeel(lafUpdateCallback)
 {
+    // a single instance of tooltip window is required and used by JUCE everywhere a tooltip is required.
     m_toolTipWindowInstance = std::make_unique<TooltipWindow>();
 
+    // create the configuration object (is being initialized from disk automatically)
     m_config = std::make_unique<AppConfiguration>(JUCEAppBasics::AppConfigurationBase::getDefaultConfigFilePath());
     m_config->addDumper(this);
-    m_config->addWatcher(this);
 
     // check if config creation was able to read a valid config from disk...
     if (!m_config->isValid())
@@ -58,22 +59,23 @@ MainSpaConBridgeComponent::MainSpaConBridgeComponent(std::function<void(DbLookAn
         m_config->triggerConfigurationDump();
     }
 
-    // enshure the config is processed and contents forwarded to already existing application components.
-    onConfigUpdated();
-    m_config->triggerWatcherUpdate();
+    // add this main component to watchers
+    m_config->addWatcher(this, true); // this initial update cannot yet reach all parts of the app, esp. settings page that relies on fully initialized pagecomponentmanager, therefor a manual watcher update is triggered below
 
     // enshure the controller singleton is created
     auto ctrl = SpaConBridge::Controller::GetInstance();
+    jassert(ctrl);
     ignoreUnused(ctrl);
-
-    // enshure the overviewmanager singleton is created
+    // enshure the pagemanager singleton is created
     auto pageMgr = SpaConBridge::PageComponentManager::GetInstance();
-    if (pageMgr)
-    {
-        // get the overview component from manager to use as central element for app ui
-        auto pageContainer = pageMgr->GetPageContainer();
-        addAndMakeVisible(pageContainer);
-    }
+    jassert(pageMgr);
+
+    // get the overview component from manager to use as central element for app ui
+    auto pageContainer = pageMgr->GetPageContainer();
+    addAndMakeVisible(pageContainer);
+
+    // do the initial update for the whole application with config contents
+    m_config->triggerWatcherUpdate();
 
     setSize(896, 414);
 }