+++ Source/PagedUI/PageComponentManager.cpp
@@ -69,6 +69,13 @@ PageComponentManager::PageComponentManager()
 
 	// Default overview window properties.
 	m_pageContainer = nullptr;
+
+	// Initialization of members to a usable default makes sense 
+	// when no config xml is available on app start and therefor 
+	// init values of this instance are going to be used for default config.
+	m_activePage = UPI_SoundObjects;
+	for (int id = UPI_InvalidMin + 1; id < UPI_InvalidMax; id++) m_enabledPages.push_back(static_cast<UIPageId>(id));
+	m_lookAndFeelType = DbLookAndFeelBase::LAFT_Dark;
 }
 
 /**
@@ -146,25 +153,26 @@ void PageComponentManager::ClosePageContainer(bool destroy)
 }
 
 /**
- * Get the currently active tab within the overview window.
- * @return The currently active tab.
+ * Getter for the currently active page of the main window.
+ * @return The currently active page.
  */
-int PageComponentManager::GetActiveTab() const
+UIPageId PageComponentManager::GetActivePage() const
 {
-	return m_selectedTab;
+	return m_activePage;
 }
 
 /**
- * Set the currently active tab within the overview window.
- * @param tabIdx	The currently active tab index.
+ * Setter for the currently active tab of the main window.
+ * @param pageIdx	The currently active page id.
+ * @param dontSendNotification	Indication if the configuration update shall be triggerd as well
  */
-void PageComponentManager::SetActiveTab(int tabIdx, bool dontSendNotification)
+void PageComponentManager::SetActivePage(UIPageId pageId, bool dontSendNotification)
 {
-	m_selectedTab = tabIdx;
+	m_activePage = pageId;
 
 	if (m_pageContainer != nullptr)
 	{
-		m_pageContainer->SetActiveTab(tabIdx);
+		m_pageContainer->SetActivePage(pageId);
 	}
 
 	if (!dontSendNotification)
@@ -174,21 +182,32 @@ void PageComponentManager::SetActiveTab(int tabIdx, bool dontSendNotification)
 }
 
 /**
- * Get the currently selected coordinate mapping used for the multi-slider.
- * @return The selected mapping area.
+ * Getter for the currently enabled pages of the main window.
+ * @return The currently enabled pages.
  */
-int PageComponentManager::GetSelectedMapping() const
+const std::vector<UIPageId>& PageComponentManager::GetEnabledPages() const
 {
-	return m_selectedMapping;
+	return m_enabledPages;
 }
 
 /**
- * Set the currently selected coordinate mapping used for the multi-slider.
- * @param mapping	The new selected mapping area.
+ * Setter for the currently enabled pages of the main window.
+ * @param enabledPages	The pages to set as currently currently enabled.
+ * @param dontSendNotification	Indication if the configuration update shall be triggerd as well
  */
-void PageComponentManager::SetSelectedMapping(int mapping)
+void PageComponentManager::SetEnabledPages(const std::vector<UIPageId>& enabledPages, bool dontSendNotification)
 {
-	m_selectedMapping = mapping;
+	m_enabledPages = enabledPages;
+
+	if (m_pageContainer != nullptr)
+	{
+		m_pageContainer->SetEnabledPages(enabledPages);
+	}
+
+	if (!dontSendNotification)
+	{
+		triggerConfigurationUpdate(false);
+	}
 }
 
 /**
@@ -327,20 +346,39 @@ void PageComponentManager::SetMatrixOutputTableCollapsed(bool collapsed)
 }
 
 /**
+ * Proxy Getter for the pinned scenes in Scenes Page.
+ * Forwards the call to PageContainerComponent.
+ * @return	The pinned scenes of Scenes Page.
+ */
+std::vector<std::pair<std::pair<int, int>, std::string>> PageComponentManager::GetScenesPagePinnedScenes()
+{
+	if (m_pageContainer)
+		return m_pageContainer->GetScenesPagePinnedScenes();
+	else
+	{
+		jassertfalse;
+		return std::vector<std::pair<std::pair<int, int>, std::string>>();
+	}
+}
+
+/**
+ * Proxy Setter for the pinned scenes in Scenes Page.
+ * Forwards the call to PageContainerComponent.
+ * @param pinnedScenes	The pinned scenes of Scenes Page.
+ */
+void PageComponentManager::SetScenesPagePinnedScenes(const std::vector<std::pair<std::pair<int, int>, std::string>>& pinnedScenes)
+{
+	if (m_pageContainer)
+		m_pageContainer->SetScenesPagePinnedScenes(pinnedScenes);
+}
+
+/**
  * Get the currently selected coordinate mapping used for the multi-slider.
  * @return The selected mapping area.
  */
 DbLookAndFeelBase::LookAndFeelType PageComponentManager::GetLookAndFeelType() const
 {
-	if (m_pageContainer != nullptr)
-	{
-		return m_pageContainer->GetLookAndFeelType();
-	}
-	else
-	{
-		jassertfalse;
-		return DbLookAndFeelBase::LAFT_InvalidFirst;
-	}
+	return m_lookAndFeelType;
 }
 
 /**
@@ -349,14 +387,11 @@ DbLookAndFeelBase::LookAndFeelType PageComponentManager::GetLookAndFeelType() co
  */
 void PageComponentManager::SetLookAndFeelType(DbLookAndFeelBase::LookAndFeelType lookAndFeelType, bool dontSendNotification)
 {
-	if (m_pageContainer != nullptr)
-	{
-		m_pageContainer->SetLookAndFeelType(lookAndFeelType);
-	}
+	m_lookAndFeelType = lookAndFeelType;
 
 	if (!dontSendNotification)
 	{
-		triggerConfigurationUpdate(false);
+		triggerConfigurationUpdate(true); // we do want to include watcher update, since only that way the whole application is set to new LAFT (onConfigUpdated in main component)
 	}
 }
 
@@ -371,7 +406,12 @@ bool PageComponentManager::setStateXml(XmlElement* stateXml)
 	if (!stateXml || (stateXml->getTagName() != AppConfiguration::getTagName(AppConfiguration::TagID::UICONFIG)))
 		return false;
 
+	if (!m_pageContainer)
+		return false;
+
 	auto retVal = true;
+
+	m_pageContainer->SetPagesBeingInitialized(true);
 
 	// Handle the look and feel type from xml first, since this is set as active dropdown index in overview. If we do not do this first, 
 	// the default selected index will we written to config due to update trigger from SetActiveTab
@@ -398,15 +438,32 @@ bool PageComponentManager::setStateXml(XmlElement* stateXml)
 		auto activeTabTextElement = activeTabXmlElement->getFirstChildElement();
 		if (activeTabTextElement && activeTabTextElement->isTextElement())
 		{
-			auto tabIdx = activeTabTextElement->getText().getIntValue();
+			auto pageId = GetPageIdFromName(activeTabTextElement->getText());
 
-			if (tabIdx != -1)
+			if (pageId > UPI_InvalidMin && pageId < UPI_InvalidMax)
 			{
-				SetActiveTab(tabIdx, true);
+				SetActivePage(pageId, true);
 			}
 			else
 				retVal = false;
 		}
+	}
+
+	auto enabledPagesXmlElement = stateXml->getChildByName(AppConfiguration::getTagName(AppConfiguration::TagID::ENABLEDPAGES));
+	if (enabledPagesXmlElement)
+	{
+		auto enabledPages = std::vector<UIPageId>();
+
+		auto enabledPagesTextElement = enabledPagesXmlElement->getFirstChildElement();
+		if (enabledPagesTextElement && enabledPagesTextElement->isTextElement())
+		{
+			auto enabledPagesIdStrings = StringArray();
+			enabledPagesIdStrings.addTokens(enabledPagesTextElement->getText(), ", ");
+			for (auto const& pageIdString : enabledPagesIdStrings)
+				enabledPages.push_back(static_cast<UIPageId>(pageIdString.getIntValue()));
+		}
+
+		SetEnabledPages(enabledPages, true);
 	}
 
 	auto soundobjectTableXmlElement = stateXml->getChildByName(AppConfiguration::getTagName(AppConfiguration::TagID::SOUNDOBJECTTABLE));
@@ -496,6 +553,33 @@ bool PageComponentManager::setStateXml(XmlElement* stateXml)
 		}
 	}
 
+	auto scenesPageXmlElement = stateXml->getChildByName(AppConfiguration::getTagName(AppConfiguration::TagID::SCENESPAGE));
+	if (scenesPageXmlElement)
+	{
+		auto pinnedScenesXmlElement = scenesPageXmlElement->getChildByName(AppConfiguration::getTagName(AppConfiguration::TagID::PINNEDSCENES));
+		if (pinnedScenesXmlElement)
+		{
+			auto pinnedScenes = std::vector<std::pair<std::pair<int, int>, std::string>>();
+			for (auto sceneXmlElement : pinnedScenesXmlElement->getChildIterator())
+			{
+				if (sceneXmlElement && sceneXmlElement->getTagName() == AppConfiguration::getTagName(AppConfiguration::TagID::SCENE))
+				{
+					auto sceneIndexMajor = sceneXmlElement->getIntAttribute(AppConfiguration::getAttributeName(AppConfiguration::AttributeID::INDEXMAJOR));
+					auto sceneIndexMinor = sceneXmlElement->getIntAttribute(AppConfiguration::getAttributeName(AppConfiguration::AttributeID::INDEXMINOR));
+					auto sceneName = String();
+					auto sceneNameTextXmlElement = sceneXmlElement->getFirstChildElement();
+					if (sceneNameTextXmlElement && sceneNameTextXmlElement->isTextElement())
+						sceneName = sceneNameTextXmlElement->getText();
+					pinnedScenes.push_back(std::make_pair(std::make_pair(sceneIndexMajor, sceneIndexMinor), sceneName.toStdString()));
+				}
+			}
+
+			SetScenesPagePinnedScenes(pinnedScenes);
+		}
+	}
+    
+    m_pageContainer->SetPagesBeingInitialized(false);
+
 	return retVal;
 }
 
@@ -512,7 +596,16 @@ std::unique_ptr<XmlElement> PageComponentManager::createStateXml()
 	{
 		auto activeTabXmlElement = uiCfgXmlElement->createNewChildElement(AppConfiguration::getTagName(AppConfiguration::TagID::ACTIVETAB));
 		if (activeTabXmlElement)
-			activeTabXmlElement->addTextElement(String(GetActiveTab()));
+			activeTabXmlElement->addTextElement(GetPageNameFromId(GetActivePage()));
+
+		auto enabledPagesXmlElement = uiCfgXmlElement->createNewChildElement(AppConfiguration::getTagName(AppConfiguration::TagID::ENABLEDPAGES));
+		if (enabledPagesXmlElement)
+		{
+			auto enabledPagesIdStrings = StringArray();
+			for (auto const& pageId : GetEnabledPages())
+				enabledPagesIdStrings.add(String(pageId));
+			enabledPagesXmlElement->addTextElement(enabledPagesIdStrings.joinIntoString(", "));
+		}
 
 		auto lookAndFeelXmlElement = uiCfgXmlElement->createNewChildElement(AppConfiguration::getTagName(AppConfiguration::TagID::LOOKANDFEELTYPE));
 		if (lookAndFeelXmlElement)
@@ -548,6 +641,26 @@ std::unique_ptr<XmlElement> PageComponentManager::createStateXml()
 			auto collapsedXmlElement = matrixOutputTableXmlElement->createNewChildElement(AppConfiguration::getTagName(AppConfiguration::TagID::COLLAPSED));
 			if (collapsedXmlElement)
 				collapsedXmlElement->addTextElement(String(IsMatrixOutputTableCollapsed() ? 1 : 0));
+		}
+
+		auto scenesPageXmlElement = uiCfgXmlElement->createNewChildElement(AppConfiguration::getTagName(AppConfiguration::TagID::SCENESPAGE));
+		if (scenesPageXmlElement)
+		{
+			auto pinnedScenesXmlElement = scenesPageXmlElement->createNewChildElement(AppConfiguration::getTagName(AppConfiguration::TagID::PINNEDSCENES));
+			if (pinnedScenesXmlElement)
+			{
+				auto pinnedScenes = GetScenesPagePinnedScenes();
+				for (auto const& pinnedScene : pinnedScenes)
+				{
+					auto pinnedSceneXmlElement = pinnedScenesXmlElement->createNewChildElement(AppConfiguration::getTagName(AppConfiguration::TagID::SCENE));
+					if (pinnedSceneXmlElement)
+					{
+						pinnedSceneXmlElement->setAttribute(AppConfiguration::getAttributeName(AppConfiguration::AttributeID::INDEXMAJOR), pinnedScene.first.first);
+						pinnedSceneXmlElement->setAttribute(AppConfiguration::getAttributeName(AppConfiguration::AttributeID::INDEXMINOR), pinnedScene.first.second);
+						pinnedSceneXmlElement->addTextElement(pinnedScene.second);
+					}
+				}
+			}
 		}
 	}