+++ Source/PagedUI/PageComponents/StatisticsPage/StatisticsPageComponent.cpp
@@ -56,7 +56,7 @@ StatisticsPageComponent::StatisticsPageComponent()
 
 	auto config = SpaConBridge::AppConfiguration::getInstance();
 	if (config)
-		config->addWatcher(this);
+		config->addWatcher(this, true);
 }
 
 /**