+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added "Pinnable" Scenes (create custom Scene Index recall buttons by "pinning" current Scene)
+- Added "General Settings" section to settings page, allowing enabling/disabling of pages to be shown on UI, and moved existing look and feel selection dropdown from bottom bar to "General Settings" section
+
+### Changed
+- Introduced major performance improvement regarding all UI interactions that affect the application xml configuration
+
+### Fixed
+- App stability and minor bugs
 
 ## [0.3.2] - 2021-05-04
 ### Added